[FIX] 메인화면 인기 봉사 total 자료형 및 봉사활동 상세보기 지원 분기처리 오류 수정

+++ app/src/main/java/com/sharehands/sharehands_frontend/network/home/HomeData.kt
@@ -36,5 +36,5 @@ data class PopularItem(
     @SerializedName("likes") val likes: Long,
     @SerializedName("scraps") val scraps: Long,
     @SerializedName("serviceId") val serviceId: Long,
-    @SerializedName("total") val total: Long
+    @SerializedName("total") val total: Double
 )
